addded tally function
