+++ rust/catalyst-contest/src/contest_ballot/payload/encrypted_block.rs
