+++ rust/catalyst-contest/src/lib.rs
@@ -5,32 +5,37 @@
 //! [documentation]: https://docs.dev.projectcatalyst.io/libs/main/architecture/08_concepts/signed_doc/docs/contest_ballot/
 
 pub mod checkpoint;
+pub mod contest_ballot;
 pub mod contest_delegation;
 pub mod contest_parameters;
 
-mod contest_ballot;
-
 use catalyst_signed_doc::providers::CatalystSignedDocumentProvider;
 
-pub use crate::contest_ballot::{
-    Choices, ContentBallotPayload, ContestBallot, ContestBallotRule, EncryptedBlock,
-    EncryptedChoices,
-};
-use crate::contest_parameters::ContestParameters;
+use crate::contest_parameters::{Choices, ContestParameters};
+
+/// Contest Tally Result type.
+#[derive(Debug, Clone)]
+pub struct TallyResult {
+    /// Contest choices, defined by the 'Contest Parameters' document
+    #[allow(dead_code)]
+    choices: Choices,
+}
 
 /// Contest tally procedure based on the provided 'Contest Parameters' document.
 /// Collects all necessary `ContestBallot`, `Proposal`, `ContestDelegation` documents
 /// which are associate with the provided `ContestParameters`.
-///
-/// Filling the provided `contest_parameters` problem report if something goes wrong.
 ///
 /// # Errors
 ///  - `provider` returns error
 pub fn tally(
     contest_parameters: &ContestParameters,
     provider: &dyn CatalystSignedDocumentProvider,
-) -> anyhow::Result<()> {
+) -> anyhow::Result<TallyResult> {
+    let res = TallyResult {
+        choices: contest_parameters.choices().clone(),
+    };
+
     let _proposals = contest_parameters.get_associated_proposals(provider)?;
 
-    Ok(())
+    Ok(res)
 }
